refactor: moved the components to a components folder
