+++ frontend/mydollarbot-frontend/src/components/AddSpendingRecord.tsx
